requirements1.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.3
defusedxml==0.7.1
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-enum-choices==2.1.4
django-filter==22.1
django-taggit==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environs==9.5.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.17.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.30
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5